optimize link commands by move out same functionality in separate file
