+++ scripts/link/postlink/android/postlink.js
@@ -1,64 +1,35 @@
+var linkTools = require('../../linkTools');
 var fs = require("fs");
-var glob = require("glob");
-var path = require("path");
 var inquirer = require('inquirer');
 
 module.exports = () => {
 
     console.log("Running android postlink script");
 
-    var ignoreFolders = { ignore: ["node_modules/**", "**/build/**"] };
-    var buildGradlePath = path.join("android", "app", "build.gradle");
-    var manifestPath = glob.sync("**/AndroidManifest.xml", ignoreFolders)[0];
-
-    function findMainApplication() {
-        if (!manifestPath) {
-            return null;
-        }
-
-        var manifest = fs.readFileSync(manifestPath, "utf8");
-
-        // Android manifest must include single 'application' element
-        var matchResult = manifest.match(/application\s+android:name\s*=\s*"(.*?)"/);
-        if (matchResult) {
-            var appName = matchResult[1];
-        } else {
-            return null;
-        }
-        
-        var nameParts = appName.split('.');
-        var searchPath = glob.sync("**/" + nameParts[nameParts.length - 1] + ".java", ignoreFolders)[0];
-        return searchPath;
-    }
-
-    var mainApplicationPath = findMainApplication() || glob.sync("**/MainApplication.java", ignoreFolders)[0];
+    var buildGradlePath = linkTools.getBuildGradlePath();
+    var mainApplicationPath = linkTools.getMainApplicationLocation();
 
     // 1. Add the getJSBundleFile override
-    var getJSBundleFileOverride = "\n\n\t\t@Override\n\t\tprotected String getJSBundleFile()" +
-    "{\n\t\t\treturn CodePush.getJSBundleFile();\n\t\t}\n";
-
-    function isAlreadyOverridden(codeContents) {
-        return /@Override\s*\n\s*protected String getJSBundleFile\(\)\s*\{[\s\S]*?\}/.test(codeContents);
-    }
+    var getJSBundleFileOverride = linkTools.getJSBundleFileOverride;
 
     if (mainApplicationPath) {
         var mainApplicationContents = fs.readFileSync(mainApplicationPath, "utf8");
-        if (isAlreadyOverridden(mainApplicationContents)) {
+        if (linkTools.isJsBundleChanged(mainApplicationContents)) {
             console.log(`"getJSBundleFile" is already overridden`);
         } else {
-            var reactNativeHostInstantiation = "new ReactNativeHost(this) {";
+            var reactNativeHostInstantiation = linkTools.reactNativeHostInstantiation;
             mainApplicationContents = mainApplicationContents.replace(reactNativeHostInstantiation,
                 `${reactNativeHostInstantiation}${getJSBundleFileOverride}`);
             fs.writeFileSync(mainApplicationPath, mainApplicationContents);
         }
     } else {
-        var mainActivityPath = glob.sync("**/MainActivity.java", ignoreFolders)[0];
+        var mainActivityPath = linkTools.getMainActivityPath();
         if (mainActivityPath) {
             var mainActivityContents = fs.readFileSync(mainActivityPath, "utf8");
-            if (isAlreadyOverridden(mainActivityContents)) {
+            if (linkTools.isJsBundleChangedy(mainActivityContents)) {
                 console.log(`"getJSBundleFile" is already overridden`);
             } else {
-                var mainActivityClassDeclaration = "public class MainActivity extends ReactActivity {";
+                var mainActivityClassDeclaration = linkTools.mainActivityClassDeclaration;
                 mainActivityContents = mainActivityContents.replace(mainActivityClassDeclaration,
                     `${mainActivityClassDeclaration}${getJSBundleFileOverride}`);
                 fs.writeFileSync(mainActivityPath, mainActivityContents);
@@ -79,22 +50,22 @@ module.exports = () => {
     // 2. Add the codepush.gradle build task definitions
     var buildGradleContents = fs.readFileSync(buildGradlePath, "utf8");
     var reactGradleLink = buildGradleContents.match(/\napply from: ["'].*?react\.gradle["']/)[0];
-    var codePushGradleLink = `apply from: "../../node_modules/react-native-code-push/android/codepush.gradle"`;
+    var codePushGradleLink = linkTools.codePushGradleLink;
     if (~buildGradleContents.indexOf(codePushGradleLink)) {
         console.log(`"codepush.gradle" is already linked in the build definition`);
     } else {
         buildGradleContents = buildGradleContents.replace(reactGradleLink,
-            `${reactGradleLink}\n${codePushGradleLink}`);
+            `${reactGradleLink}${codePushGradleLink}`);
         fs.writeFileSync(buildGradlePath, buildGradleContents);
     }
 
     //3. Add deployment key
-    var stringsResourcesPath = glob.sync("**/strings.xml", ignoreFolders)[0];
+    var stringsResourcesPath = linkTools.getStringsResourcesPath();
     if (!stringsResourcesPath) {
         return Promise.reject(new Error(`Couldn't find strings.xml. You might need to update it manually.`));
     } else {
         var stringsResourcesContent = fs.readFileSync(stringsResourcesPath, "utf8");
-        var deploymentKeyName = "reactNativeCodePush_androidDeploymentKey";
+        var deploymentKeyName = linkTools.deploymentKeyName;
         if (~stringsResourcesContent.indexOf(deploymentKeyName)) {
             console.log(`${deploymentKeyName} already specified in the strings.xml`);
         } else {
